Hydra.BehaviorSpec

Documentation

spec :: Spec

waitUntil :: (HasCallStack, MonadThrow m, MonadAsync m, MonadTimer m, IsChainState tx) => [TestHydraNode tx m] -> ServerOutput tx -> m ()

Wait for some output at some node(s) to be produced eventually. See waitUntilMatch for how long it waits.

waitUntilMatch :: (HasCallStack, MonadThrow m, MonadAsync m, MonadTimer m) => [TestHydraNode tx m] -> (ServerOutput tx -> Bool) -> m ()

Wait for some output to match some predicate eventually. This will not wait forever, but for a VERY long time (1000 years) to get a nice error location. Should not be an issue when used within shouldRunInSim.

waitMatch :: MonadThrow m => TestHydraNode tx m -> (ServerOutput tx -> Maybe a) -> m a

Wait for an output matching the predicate and extracting some value. This will loop forever until a match has been found.

data TestHydraNode tx m

A thin layer around HydraNode to be able to waitFor.

Constructors

TestHydraNode

Fields

send :: ClientInput tx -> m ()

waitForNext :: m (ServerOutput tx)

injectChainEvent :: ChainEvent tx -> m ()

serverOutputs :: m [ServerOutput tx]

data ConnectToChain tx m

A simulated chain that just echoes PostChainTx as Observations of OnChainTx onto all connected nodes. It can also $sel:rollbackAndForward:ConnectToChain any number of these "transactions".

Constructors

ConnectToChain

Fields

chainComponent :: HydraNode tx m -> m (HydraNode tx m)

tickThread :: Async m ()

rollbackAndForward :: Natural -> m ()

withSimulatedChainAndNetwork :: (MonadSTM m, MonadTime m, MonadDelay m, MonadAsync m) => (ConnectToChain SimpleTx m -> m ()) -> m ()

With-pattern wrapper around simulatedChainAndNetwork which does cancel the $sel:tickThread:ConnectToChain. Also, this will fix tx to SimpleTx so that it can pick an initial chain state to play back to our test nodes.

class IsChainState a => IsChainStateTest a where

Class to manipulate the chain state by advancing it's slot in simulatedChainAndNetwork.

Methods

advanceSlot :: ChainStateType a -> ChainStateType a

simulatedChainAndNetwork :: (MonadSTM m, MonadTime m, MonadDelay m, MonadAsync m, IsChainStateTest tx) => ChainStateType tx -> m (ConnectToChain tx m)

Creates a simulated chain and network by returning a handle with a HydraNode decorator to connect it to the simulated chain. NOTE: The $sel:tickThread:ConnectToChain needs to be canceled after use. Use withSimulatedChainAndNetwork instead where possible.

toOnChainTx :: UTCTime -> PostChainTx tx -> OnChainTx tx

Derive an OnChainTx from PostChainTx to simulate a "perfect" chain. NOTE(SN): This implementation does *NOT* honor the HeadParameters and announces hard-coded contestationDeadlines.

testContestationPeriod :: ContestationPeriod

nothingHappensFor :: (MonadTimer m, MonadThrow m, IsChainState tx) => TestHydraNode tx m -> DiffTime -> m ()

withHydraNode :: forall s a. SigningKey HydraKey -> [Party] -> ConnectToChain SimpleTx (IOSim s) -> (TestHydraNode SimpleTx (IOSim s) -> IOSim s a) -> IOSim s a

createTestHydraNode :: MonadSTM m => TQueue m (ServerOutput tx) -> TVar m [ServerOutput tx] -> HydraNode tx m -> TestHydraNode tx m

createHydraNode :: (MonadDelay m, MonadAsync m) => Ledger tx -> ChainStateType tx -> SigningKey HydraKey -> [Party] -> TQueue m (ServerOutput tx) -> TVar m [ServerOutput tx] -> ConnectToChain tx m -> m (HydraNode tx m)

openHead :: TestHydraNode SimpleTx (IOSim s) -> TestHydraNode SimpleTx (IOSim s) -> IOSim s ()

matchFanout :: PostChainTx tx -> Bool

assertHeadIsClosed :: (HasCallStack, MonadThrow m) => ServerOutput tx -> m ()

assertHeadIsClosedWith :: (HasCallStack, MonadThrow m) => SnapshotNumber -> ServerOutput tx -> m ()