Hydra.Chain.Direct.Contract.Abort

Description

Mutation-based script validator tests for the abort transaction where a healthyAbortTx gets mutated by an arbitrary AbortMutation.

Documentation

healthyAbortTx :: HasCallStack => (Tx, UTxO)

healthyHeadInput :: TxIn

healthyHeadParameters :: HeadParameters

healthyInitials :: [UTxOWithScript]

healthyCommits :: [UTxOWithScript]

thereIsAtLeastOneCommit :: ([UTxOWithScript], [UTxOWithScript]) -> Bool

healthyParties :: [Party]

propHasInitial :: (Tx, UTxO) -> Property

propHasCommit :: (Tx, UTxO) -> Property

data AbortMutation

Constructors

MutateParties

DropOneCommitOutput

MutateHeadScriptInput

BurnOneTokenMore

MutateThreadTokenQuantity
Meant to test that the minting policy is burning all PTs present in tx

DropCollectedInput

MutateRequiredSigner

MutateHeadId
Simply change the currency symbol of the ST.

UseInputFromOtherHead

genAbortMutation :: (Tx, UTxO) -> Gen SomeMutation