Hydra.Chain.Direct.Contract.CollectCom

Documentation

healthyCollectComTx :: (Tx, UTxO)

healthyParticipants :: [VerificationKey PaymentKey]

healthyCommits :: Map TxIn HealthyCommit

healthyContestationPeriod :: ContestationPeriod

healthyHeadTxIn :: TxIn

healthyHeadTxOut :: TxOut CtxUTxO

healthyCollectComInitialDatum :: State

healthyOnChainParties :: [Party]

healthyParties :: [Party]

data HealthyCommit

Constructors

HealthyCommit

Fields

participant :: OnChainId

txOut :: TxOut CtxUTxO

committed :: UTxO

healthyCommitOutput :: OnChainId -> Party -> UTxO -> (TxIn, HealthyCommit)

data CollectComMutation

Constructors

NotContinueContract
Ensures collectCom does not allow any output address but νHead.

ExtractSomeValue
Needs to prevent that not all value is collected into the head output.

MutateOpenUTxOHash

MutateCommitToInitial
Ensures collectCom cannot collect from an initial UTxO.

MutateNumberOfParties
Every party should have commited and been taken into account for the collectCom transaction to be valid. Here we increase the number of parties in input and output but keep the commits unchanged. This simulates the situation where one participant would not have commited already or whose commit would have been ignored by the collectCom transaction.

MutateHeadId

MutateRequiredSigner

MutateTokenMintingOrBurning
Minting or burning of tokens should not be possible in collectCom.

RemoveSTFromOutput
νCommit validator checks the ST is in the output

genCollectComMutation :: (Tx, UTxO) -> Gen SomeMutation