Hydra.Chain.Direct.Contract.Contest

Documentation

healthyContestTx :: (Tx, UTxO)

addParticipationTokens :: [Party] -> TxOut CtxUTxO -> TxOut CtxUTxO

healthyContestSnapshot :: Snapshot Tx

healthyContestSnapshotNumber :: SnapshotNumber

healthyContestUTxO :: UTxO

healthyContestUTxOHash :: BuiltinByteString

healthyClosedState :: State

healthySlotNo :: SlotNo

healthyContestationDeadline :: POSIXTime

healthyContestationPeriodSeconds :: Integer

healthyClosedSnapshotNumber :: SnapshotNumber

healthyClosedUTxOHash :: BuiltinByteString

healthyClosedUTxO :: UTxO

somePartyCardanoVerificationKey :: VerificationKey PaymentKey

healthySigningKeys :: [SigningKey]

healthyParties :: [Party]

healthyOnChainParties :: [Party]

healthySignature :: SnapshotNumber -> MultiSignature (Snapshot Tx)

data ContestMutation

Constructors

MutateSignatureButNotSnapshotNumber
Ensure signatures are actually checked.

MutateToNonNewerSnapshot
Ensure too old snapshot are not valid.

MutateRequiredSigner
Ensure that it's performed by a Head party

MutateContestUTxOHash
Ensure output state is consistent with redeemer

MutateParties
Change parties stored in the state, causing multisig to fail

MutateValidityPastDeadline
Change the validity interval of the transaction to a value greater than the contestation deadline

genContestMutation :: (Tx, UTxO) -> Gen SomeMutation