Hydra.Chain.Direct.Contract.Mutation

Description

Provides building blocks for Mutation testing of Contracts.

Introduction

Traditional Mutation testing is a testing technique that introduces small modifications like changing a comparison operator, or modifying constants, into a program and checks whether or not the existing tests "kill" the produced mutants, eg. fail. Mutation testing requires somewhat complex tooling because it needs to modify the source code, in limited and semantically meaningful ways in order to generate code that won't be rejected by the compiler.

Recall that Plutus eUTxO validators are boolean expressions of the form:

validator : Datum -> Redeemer -> ScriptContext -> Bool

All things being equal, "mutating" a validator so that it returns False instead of True can be done:

Either by mutating the code of the validator implementation,

Or by mutating its arguments.

This simple idea lead to the following strategy to test-drive validator scripts:

Start with a validator that always return True,

Write a positive property test checking valid transactions are accepted by the validator(s),

Write a negative property test checking invalid transactions are rejected. This is where mutations are introduced, each different mutation type representing some possible "attack",

Watch one or the other properties fail and enhance the validators code to make them pass,

Rinse and repeat.

Generic Property and Mutations

Given a transaction with some UTxO context, and a function that generates SomeMutation from a valid transaction and context pair, the generic propMutation property checks applying any generated mutation makes the mutated (hence expectedly invalid) transaction fail the validation stage.

propMutation :: (Tx, Utxo) -> ((Tx, Utxo) -> Gen SomeMutation) -> Property
propMutation (tx, utxo) genMutation =
  forAll _ Property (genMutation (tx, utxo)) $ SomeMutation{label, mutation} ->
    (tx, utxo)
      & applyMutation mutation
      & propTransactionDoesNotValidate

To this basic property definition we add a checkCoverage that ensures the set of generated mutations covers a statistically significant share of each of the various possible mutations classified by their label.

The SomeMutation type is simply a wrapper that attaches a label to a proper Mutation which is the interesting bit here.

The Mutation type enumerates various possible "atomic" mutations which preserve the structural correctness of the transaction but should make a validator fail.

data Mutation
  = ChangeHeadRedeemer Head.Input
  | ChangeHeadDatum Head.State
  ...
  | Changes [Mutation]

The constructors should hopefully be self-explaining but for the last one. Some interesting mutations we want to make require more than one "atomic" change to represent a possible validator failure. For example, we wanted to check that the Commit validator, in the context of a CollectCom transaction, verifies the state (Input) of the Head validator is correct. But to be interesting, this mutation needs to ensure the transition verified by the Head state machine is valid, which requires changing both the datum and the redeemer of the consumed head output.

Transaction-specific Mutations

To be run the propMutation requires a starting "healthy" (valid) transaction and a specialised generating function. It is instantiated in the test runner by providing these two elements. For example, the ContractSpec module has the following property check:

describe CollectCom $ do
  prop "does not survive random adversarial mutations" $
    propMutation healthyCollectComTx genCollectComMutation

The interesting part is the genCollectComMutation (details of the Mutation generators are omitted):

genCollectComMutation :: (Tx, Utxo) -> Gen SomeMutation
genCollectComMutation (tx, utxo) =
  oneof
    [ SomeMutation MutateOpenOutputValue . ChangeOutput ...
    , SomeMutation MutateOpenUtxoHash . ChangeOutput ...
    , SomeMutation MutateHeadScriptInput . ChangeInput ...
    , SomeMutation MutateHeadTransition $ do
        changeRedeemer ChangeHeadRedeemer <$ ...
        changeDatum ChangeHeadDatum <$ ...
        pure $ Changes [changeRedeemer, changeDatum]
    ]

Here we have defined four different type of mutations that are interesting for the CollectCom transaction and represent possible attack vectors:

Changing the Head output's value, which would imply some of the committed funds could be "stolen" by the party posting the transaction,

Tampering with the content of the UTxO committed to the Head,

Trying to collect commits without running the Head validator,

Trying to collect commits in another Head state machine transition.

Running Properties

When such a property test succeeds we get the following report which shows the distribution of the various mutations that were tested.

Hydra.Chain.Direct.Contract
  CollectCom
    does not survive random adversarial mutations
      +++ OK, passed 200 tests.

      CollectComMutation (200 in total):
      30.5% MutateOpenUtxoHash
      27.0% MutateHeadTransition
      23.5% MutateOpenOutputValue
      19.0% MutateHeadScriptInput

Finished in 18.1146 seconds

In the case of a failure we get a detailed report on the context of the failure.

Properties

propMutationOnChain :: (Tx, UTxO) -> ((Tx, UTxO) -> Gen SomeMutation) -> Property

A Property checking a mutation is not validated. This property takes an initial (transaction, UTxO) pair that is supposedly valid, passes it to a generator that produces some mutations, then assert the resulting (transaction', UTxO') pair fails the validation process.

Note that only "level 2" validation is run, e.g the transaction is assume to be structurally valid and having passed "level 1" checks.

propMutationOffChain :: (Tx, UTxO) -> ((Tx, UTxO) -> Gen SomeMutation) -> Gen SomeOnChainHeadState -> Property

propTransactionDoesNotValidate :: (Tx, UTxO) -> Property

A Property checking some (transaction, UTxO) pair is invalid.

propTransactionValidates :: (Tx, UTxO) -> Property

A Property checking some (transaction, UTxO) pair is valid.

propTransactionIsNotObserved :: (Tx, UTxO) -> SomeOnChainHeadState -> Property

A Property checking some (on-chain valid) (transaction, UTxO) is not properly observe given a configuration.

Mutations

data SomeMutation

Existential wrapper SomeMutation and some label type. This type is useful to provide a "generic" classification of mutation that is controlled by some custom type. The $sel:label:SomeMutation field can be passed to the genericCoverTable function to construct and display a coverage table showing the percentage of each mutation that's been applied and ensure significant coverage of all possible mutations using checkCoverage.

Constructors

forall lbl.(Typeable lbl, Enum lbl, Bounded lbl, Show lbl) => SomeMutation

Fields

label :: lbl

mutation :: Mutation

data Mutation

Basic mutations

Constructors

ChangeHeadRedeemer Input
Changes the Head script's redeemer to the given value.

ChangeHeadDatum State
Changes the Head script's datum to the given value. This modifies both the DatumHash in the UTxO context and the map of DatumHash to Datum in the transaction's witnesses.

PrependOutput (TxOut CtxTx)
Adds given output to the transaction's outputs.

RemoveOutput Word
Removes given output from the transaction's outputs.

RemoveInput TxIn
Drops the given input from the transaction's inputs

ChangeInput TxIn (TxOut CtxUTxO) (Maybe ScriptData)
Change an input's TxOut to something else. This mutation alters the redeemers of the transaction to ensure any matching redeemer for given input matches the new redeemer, otherwise the transaction would be invalid for the wrong reason (unused redeemer).

This expects Nothing if the new input is not locked by any script, and it expects Just with some potentially new redeemer if locked by a script.

ChangeOutput Word (TxOut CtxTx)
Change the transaction's output at given index to something else.

ChangeMintedValue Value
Change the transaction's minted values if it is actually minting something.

ChangeRequiredSigners [Hash PaymentKey]
Change required signers on a transaction'

ChangeValidityInterval (TxValidityLowerBound, TxValidityUpperBound)
Change the validity interval of the transaction.

Changes [Mutation]
Applies several mutations as a single atomic Mutation. This is useful to enable specific mutations that require consistent change of more than one thing in the transaction and/or UTxO set, for example to change consistently the Head script's redeemer and datum.

applyMutation :: Mutation -> (Tx, UTxO) -> (Tx, UTxO)

Apply a single Mutation to the given (transaction, UTxO) pair. ''NOTE'': This function is partial, it can raise error when some preconditions are not met by the transaction or UTxO set, for example if there's no Head script input or no datums in the transaction.

genBytes :: Gen ByteString

genHash :: Gen ByteString

Orphans

Helpers

isHeadOutput :: TxOut CtxUTxO -> Bool

Identify Head script's output.

addDatum :: TxOutDatum CtxTx -> TxBodyScriptData -> TxBodyScriptData

Adds given Datum and corresponding hash to the transaction's scripts. TODO: As we are creating the TxOutDatum from a known datum, passing a TxOutDatum is pointless and requires more work than needed to check impossible variants.

changeHeadOutputDatum :: (State -> State) -> TxOut CtxTx -> TxOut CtxTx

ensureDatums :: [TxOut CtxTx] -> TxBodyScriptData -> TxBodyScriptData

Ensures the included datums of given TxOuts are included in the transactions' TxBodyScriptData.

alterRedeemers :: (RdmrPtr -> (Data LedgerEra, ExUnits) -> (Data LedgerEra, ExUnits)) -> TxBodyScriptData -> TxBodyScriptData

Alter a transaction's redeemers map given some mapping function.

alterRedeemerFor :: Set (TxIn a) -> TxIn a -> (ScriptData -> Maybe ScriptData) -> TxBodyScriptData -> TxBodyScriptData

Remove redeemer for given TxIn from the transaction's redeemers map.

alterTxIns :: ([TxIn] -> [TxIn]) -> Tx -> Tx

alterTxOuts :: ([TxOut CtxTx] -> [TxOut CtxTx]) -> Tx -> Tx

Apply some mapping function over a transaction's outputs.

anyPayToPubKeyTxOut :: Gen (TxOut ctx)

Generates an output that pays to some arbitrary pubkey.

headTxIn :: UTxO -> TxIn

Finds the Head script's input in given UTxO set. ''NOTE'': This function is partial, it assumes the UTxO set contains a Head script output.

changeMintedValueQuantityFrom :: Tx -> Integer -> Gen Mutation

A Mutation that changes the minted/burnt quantity of all tokens.

addPTWithQuantity :: Tx -> Quantity -> Gen Mutation

A Mutation that adds an Arbitrary participation token with some quantity. As usual the quantity can be positive for minting, or negative for burning.

Orphan instances

Eq Input

Instance details

Methods

(==) :: Input -> Input -> Bool

(/=) :: Input -> Input -> Bool

Arbitrary Input

Instance details

Methods

arbitrary :: Gen Input

shrink :: Input -> [Input]

Arbitrary State

Instance details

Methods

arbitrary :: Gen State

shrink :: State -> [State]