Hydra.Chain.Direct.HandlersSpec

Documentation

genTimeHandleWithSlotInsideHorizon :: Gen (TimeHandle, SlotNo)

genTimeHandleWithSlotPastHorizon :: Gen (TimeHandle, SlotNo)

spec :: Spec

recordEventsHandler :: ChainContext -> ChainStateAt -> GetTimeHandle IO -> IO (ChainSyncHandler IO, IO [ChainEvent Tx])

Create a chain sync handler which records events as they are called back. NOTE: This ChainSyncHandler does not handle chain state updates, but uses the given ChainState constantly.

pickBlind :: Monad m => Gen a -> PropertyM m a

Like pick but using forAllBlind under the hood.

withCounterExample :: [Block] -> TVar IO ChainStateAt -> IO a -> PropertyM IO a

genBlockAt :: SlotNo -> [Tx] -> Gen Block

genRollbackPoint :: [Block] -> Gen (ChainSlot, Point Block)

Pick a block point in a list of blocks and return it along with the corresponding ChainSlot.

genSequenceOfObservableBlocks :: Gen (ChainContext, ChainStateAt, [Block])

Generate a non-sparse sequence of blocks each containing an observable transaction, starting from the returned on-chain head state.

Note that this does not generate the entire spectrum of observable transactions in Hydra, but only init and commits, which is already sufficient to observe at least one state transition and different levels of rollback.

showRollbackInfo :: (Word, Point Block) -> String