Hydra.Chain.Direct.StateSpec

Documentation

spec :: Spec

withCounterExample :: [Block] -> TVar IO SomeOnChainHeadStateAt -> IO a -> PropertyM IO a

genRollbackPoint :: [Block] -> Gen (Word, Point Block)

genSequenceOfObservableBlocks :: Gen (SomeOnChainHeadStateAt, [Block])

Generate a non-sparse sequence of blocks each containing an observable transaction, starting from the returned on-chain head state.

Note that this does not generate the entire spectrum of observable transactions in Hydra, but only init and commits, which is already sufficient to observe at least one state transition and different levels of rollback.

stAtGenesis :: SomeOnChainHeadState -> SomeOnChainHeadStateAt

propBelowSizeLimit :: forall st. Natural -> ((OnChainHeadState st -> Tx -> Property) -> Property) -> SpecWith ()

propIsValid :: forall st. ((OnChainHeadState st -> Tx -> Property) -> Property) -> SpecWith ()

forAllSt :: Testable property => (forall st. HasTransition st => OnChainHeadState st -> Tx -> property) -> Property

forAllInit :: Testable property => (OnChainHeadState 'StIdle -> Tx -> property) -> Property

forAllCommit :: Testable property => (OnChainHeadState 'StInitialized -> Tx -> property) -> Property

forAllNonEmptyByronCommit :: (PostTxError Tx -> Property) -> Property

forAllAbort :: Testable property => (OnChainHeadState 'StInitialized -> Tx -> property) -> Property

forAllCollectCom :: Testable property => (OnChainHeadState 'StInitialized -> Tx -> property) -> Property

forAllClose :: Testable property => (OnChainHeadState 'StOpen -> Tx -> property) -> Property

forAllContest :: Testable property => (OnChainHeadState 'StClosed -> Tx -> property) -> Property

forAllFanout :: Testable property => (OnChainHeadState 'StClosed -> Tx -> property) -> Property

genByronCommit :: Gen UTxO

genBlockAt :: SlotNo -> [Tx] -> Gen Block

allTransitions :: [Transition]

data Transition where

Constructors

Transition :: forall (st :: HeadStateKind). (HeadStateKindVal st, Typeable st) => TransitionFrom st -> Transition

showRollbackInfo :: (Word, Point Block) -> String

showStateRecordedAt :: SomeOnChainHeadStateAt -> Text