Hydra.HeadLogicSpec

Description

Unit tests of the the protocol logic in HeadLogic. These are very fine grained and specific to individual steps in the protocol. More high-level of the protocol logic, especially between multiple parties can be found in BehaviorSpec.

Documentation

spec :: Spec

chainEffect :: PostChainTx SimpleTx -> Effect SimpleTx

Create a chain effect with fixed chain state and slot.

observationEvent :: OnChainTx SimpleTx -> Event SimpleTx

Create an observation event with fixed chain state and slot.

hasEffect :: (HasCallStack, IsChainState tx) => Outcome tx -> Effect tx -> IO ()

hasEffectSatisfying :: (HasCallStack, IsChainState tx) => Outcome tx -> (Effect tx -> Bool) -> IO (HeadState tx)

hasNoEffectSatisfying :: (HasCallStack, IsChainState tx) => Outcome tx -> (Effect tx -> Bool) -> IO ()

isReqSn :: Effect tx -> Bool

isAckSn :: Effect tx -> Bool

inInitialState :: [Party] -> HeadState SimpleTx

inOpenState :: [Party] -> Ledger SimpleTx -> HeadState SimpleTx

inOpenState' :: [Party] -> CoordinatedHeadState SimpleTx -> HeadState SimpleTx

inClosedState :: [Party] -> HeadState SimpleTx

inClosedState' :: [Party] -> ConfirmedSnapshot SimpleTx -> HeadState SimpleTx

getConfirmedSnapshot :: HeadState tx -> Maybe (Snapshot tx)

assertNewState :: IsChainState tx => Outcome tx -> IO (HeadState tx)

assertOnlyEffects :: IsChainState tx => Outcome tx -> IO ()

applyEvent :: IsChainState tx => (HeadState tx -> Event tx -> Outcome tx) -> Event tx -> StateT (HeadState tx) IO ()

assertStateUnchangedFrom :: IsChainState tx => HeadState tx -> Outcome tx -> Expectation