Hydra.Model.MockChain

Documentation

mockChainAndNetwork :: forall m. (MonadSTM m, MonadTimer m, MonadThrow m, MonadAsync m, MonadThrow (STM m), MonadLabelledSTM m) => Tracer m DirectChainLog -> [(SigningKey HydraKey, CardanoSigningKey)] -> [Party] -> TVar m [MockHydraNode m] -> ContestationPeriod -> m (ConnectToChain Tx m, Async m ())

Provide the logic to connect a list of MockHydraNode through a dummy chain.

findOwnCardanoKey :: Party -> [(SigningKey HydraKey, CardanoSigningKey)] -> (VerificationKey PaymentKey, [VerificationKey PaymentKey])

Find Cardano vkey corresponding to our Hydra vkey using signing keys lookup. This is a bit cumbersome and a tribute to the fact the HydraNode itself has no direct knowlege of the cardano keys which are stored only at the ChainComponent level.

mkBlock :: ValidatedTx LedgerEra -> Block

createMockNetwork :: MonadSTM m => HydraNode Tx m -> TVar m [MockHydraNode m] -> Network m (Message Tx)

data MockHydraNode m

Constructors

MockHydraNode

Fields

node :: HydraNode Tx m

chainHandler :: ChainSyncHandler m

createMockChain :: (MonadTimer m, MonadThrow (STM m)) => Tracer m DirectChainLog -> ChainContext -> SubmitTx m -> m TimeHandle -> TxIn -> Chain Tx m

mkMockTxIn :: VerificationKey PaymentKey -> Word -> TxIn