Hydra.TUISpec

Documentation

spec :: Spec

gracePeriod :: NominalDiffTime

someTime :: NominalDiffTime

setupNodeAndTUI :: (TUITest -> IO ()) -> IO ()

data TUITest

Constructors

TUITest

Fields

buildVty :: IO Vty

sendInputEvent :: Event -> IO ()

getPicture :: IO ByteString

shouldRender :: HasCallStack => ByteString -> Expectation

Assert that some bytes are present in the frame. The unescaped image data is used in this assertion. That means, you do not need to include color switching escape codes etc. in your expected bytes.

withTUITest :: DisplayRegion -> (TUITest -> Expectation) -> Expectation

data TUILog

Constructors

FromCardano NodeLog

FromHydra EndToEndLog