Plutus.Codec.CBOR.EncodingSpec

Documentation

spec :: Spec

propCompareWithOracle :: Show x => (x -> Encoding) -> (x -> Encoding) -> x -> Property

Compare encoding a value x with our own encoder and a reference implementation. Counterexamples shows both encoded values, but in a pretty / readable form for the reference implementation.

data SomeValue where

Constructors

SomeValue :: forall a. Show a => a -> (a -> [a]) -> (a -> Encoding) -> (a -> Encoding) -> SomeValue

shrinkSomeValue :: SomeValue -> [SomeValue]

genSomeValue :: Gen SomeValue

genInteger :: Gen Integer

genByteString :: Gen ByteString

shrinkByteString :: ByteString -> [ByteString]

genList :: Gen a -> Gen [a]

genMap :: Gen k -> Gen v -> Gen [(k, v)]