Test.Util

Documentation

shouldRunInSim :: (forall s. IOSim s a) -> IO a

Run given action in IOSim and rethrow any exceptions.

printTrace :: forall log a. (Typeable log, ToJSON log) => Proxy log -> SimTrace a -> Text

Utility function to dump logs given a SimTrace.

shouldBe :: (HasCallStack, MonadThrow m, Eq a, Show a) => a -> a -> m ()

Lifted variant of Hspec's shouldBe.

shouldReturn :: (HasCallStack, MonadThrow m, Eq a, Show a) => m a -> a -> m ()

Lifted variant of Hspec's shouldReturn.

shouldSatisfy :: (HasCallStack, MonadThrow m, Show a) => a -> (a -> Bool) -> m ()

Lifted variant of Hspec's shouldSatisfy.

shouldNotBe :: (HasCallStack, MonadThrow m, Eq a, Show a) => a -> a -> m ()

Lifted variant of Hspec's shouldNotBe.

shouldContain :: (HasCallStack, MonadThrow m, Eq a, Show a) => [a] -> [a] -> m ()

Lifted variant of Hspec's shouldContain.

traceInIOSim :: Typeable a => Tracer (IOSim s) a

A Tracer that works in IOSim monad. This tracer uses the Output event which uses converts value traced to Dynamic which requires Typeable constraint. To retrieve the trace use selectTraceEventsDynamic applied to the correct type.