Hydra.BehaviorSpec

Documentation

spec :: Spec

waitUntil :: (HasCallStack, MonadThrow m, IsTx tx, MonadAsync m, MonadTimer m) => [TestHydraNode tx m] -> ServerOutput tx -> m ()

Wait for some output at some node(s) to be produced eventually. See waitUntilMatch for how long it waits.

waitUntilMatch :: (HasCallStack, MonadThrow m, MonadAsync m, MonadTimer m) => [TestHydraNode tx m] -> (ServerOutput tx -> Bool) -> m ()

Wait for some output to match some predicate eventually. This will not wait forever, but for a VERY long time (1000 years) to get a nice error location. Should not be an issue when used within shouldRunInSim.

data TestHydraNode tx m

A thin layer around HydraNode to be able to waitFor.

Constructors

TestHydraNode

Fields

send :: ClientInput tx -> m ()

chainEvent :: ChainEvent tx -> m ()

waitForNext :: m (ServerOutput tx)

serverOutputs :: m [ServerOutput tx]

data ConnectToChain tx m

Constructors

ConnectToChain

Fields

chainComponent :: HydraNode tx m -> m (HydraNode tx m)

history :: TVar m [PostChainTx tx]

simulatedChainAndNetwork :: MonadSTM m => m (ConnectToChain tx m)

Creates a simulated chain and network by returning a function to "monkey patch" a HydraNode such that it is connected. This is necessary, to get to know all nodes which use this function and simulate network and chain messages being sent around.

toOnChainTx :: PostChainTx tx -> ChainEvent tx

Derive an OnChainTx from PostChainTx to simulate a "perfect" chain. NOTE(SN): This implementation does *NOT* honor the HeadParameters and announces hard-coded contestationDeadlines.

testContestationPeriod :: ContestationPeriod

nothingHappensFor :: (MonadTimer m, MonadThrow m, IsTx tx) => TestHydraNode tx m -> DiffTime -> m ()

withHydraNode :: forall s a. SigningKey -> [Party] -> ConnectToChain SimpleTx (IOSim s) -> (TestHydraNode SimpleTx (IOSim s) -> IOSim s a) -> IOSim s a

createTestHydraNode :: (MonadSTM m, MonadThrow m) => TQueue m (ServerOutput tx) -> TVar m [ServerOutput tx] -> HydraNode tx m -> ConnectToChain tx m -> TestHydraNode tx m

createHydraNode :: (MonadDelay m, MonadAsync m) => Ledger tx -> SigningKey -> [Party] -> TQueue m (ServerOutput tx) -> TVar m [ServerOutput tx] -> ConnectToChain tx m -> m (HydraNode tx m)

openHead :: TestHydraNode SimpleTx (IOSim s) -> TestHydraNode SimpleTx (IOSim s) -> IOSim s ()

matchFanout :: PostChainTx tx -> Bool

assertHeadIsClosed :: (HasCallStack, MonadThrow m) => ServerOutput tx -> m ()

assertHeadIsClosedWith :: (HasCallStack, MonadThrow m) => SnapshotNumber -> ServerOutput tx -> m ()