Hydra.HeadLogicSpec

Description

Unit tests of the the protocol logic in HeadLogic. These are very fine grained and specific to individual steps in the protocol. More high-level of the protocol logic, especially between multiple parties can be found in BehaviorSpec.

Documentation

spec :: Spec

hasEffect :: (HasCallStack, IsTx tx) => Outcome tx -> Effect tx -> IO (HeadState tx)

hasEffect_ :: (HasCallStack, IsTx tx) => Outcome tx -> Effect tx -> IO ()

hasEffectSatisfying :: (HasCallStack, IsTx tx) => Outcome tx -> (Effect tx -> Bool) -> IO (HeadState tx)

hasNoEffectSatisfying :: (HasCallStack, IsTx tx) => Outcome tx -> (Effect tx -> Bool) -> IO ()

isReqSn :: Effect tx -> Bool

isAckSn :: Effect tx -> Bool

inInitialState :: [Party] -> HeadState SimpleTx

inOpenState :: [Party] -> Ledger tx -> HeadState tx

inOpenState' :: [Party] -> CoordinatedHeadState tx -> HeadState tx

inClosedState :: [Party] -> HeadState SimpleTx

inClosedState' :: [Party] -> ConfirmedSnapshot SimpleTx -> HeadState SimpleTx

getConfirmedSnapshot :: HeadState tx -> Maybe (Snapshot tx)

assertNewState :: IsTx tx => Outcome tx -> IO (HeadState tx)

applyEvent :: IsTx tx => (HeadState tx -> Event tx -> Outcome tx) -> Event tx -> StateT (HeadState tx) IO ()

assertStateUnchangedFrom :: IsTx tx => HeadState tx -> Outcome tx -> Expectation