Hydra.Model

Description

A Model of the Hydra head Protocol.

This model integrates in a single state-machine like abstraction the whole behaviour of a Hydra Head, taking into account both on-chain state and contracts, and off-chain interactions. It is written from the point of view of a pre-defined set of Hydra node operators that want to create a channel between them. It's a "happy path" model that does not implement any kind of adversarial behaviour and whose transactions are very simple: Each tx is a payment of one Ada-only UTxO transferred to another party in full, without any change.

More intricate and specialised models shall be developed once we get a firmer grasp of the whole framework, injecting faults, taking into account more parts of the stack (eg. Direct chain component), modelling more complex transactions schemes...

*NOTE**: This is still pretty much a work-in-progress esp. regarding the execution model as we explore different ways of putting this at work.

Documentation

data GlobalState

Global state of the Head protocol. While each participant in the Hydra Head protocol has its own private view of the state, we model the expected global state whose properties stem from the consensus built into the Head protocol. In other words, this state is what each node's local state should be eventually.

Constructors

Start
Start of the "world". This state is left implicit in the node's logic as it represents that state where the node does not even exist.

Idle

Fields

idleParties :: [Party]

cardanoKeys :: [VerificationKey PaymentKey]

Initial

Fields

headParameters :: HeadParameters

commits :: Committed Payment

pendingCommits :: PendingCommits

Open

Fields

headParameters :: HeadParameters

offChainState :: OffChainState

Final

Fields

finalUTxO :: UTxOType Payment

isInitialState :: GlobalState -> Bool

isFinalState :: GlobalState -> Bool

isIdleState :: GlobalState -> Bool

isPendingCommitFrom :: Party -> GlobalState -> Bool

data OffChainState

Constructors

OffChainState

Fields

confirmedUTxO :: UTxOType Payment

seenTransactions :: [Payment]

data WorldState

State maintained by the model.

Constructors

WorldState

Fields

hydraParties :: [(SigningKey HydraKey, CardanoSigningKey)]

List of parties identified by both signing keys required to run protocol. This list must not contain any duplicated key.

hydraState :: GlobalState

Expected consensus state All nodes should be in the same state.

data Nodes m

Concrete state needed to run actions against the implementation.

Constructors

Nodes

Fields

nodes :: Map Party (TestHydraNode Tx m)

Map from party identifiers to a handle for interacting with a node.

logger :: Tracer m (HydraNodeLog Tx)

Logger used by each node. The reason we put this here is because the concrete value needs to be instantiated upon the test run initialisation, outiside of the model.

newtype CardanoSigningKey

Constructors

CardanoSigningKey

Fields

signingKey :: SigningKey PaymentKey

data Payment

A single Ada-payment only transaction in our model.

Constructors

Payment

Fields

from :: CardanoSigningKey

to :: CardanoSigningKey

value :: Value

applyTx :: UTxOType Payment -> Payment -> UTxOType Payment

type ActualCommitted = UTxOType Payment

Running the model

runModel :: (MonadAsync m, MonadCatch m, MonadTimer m, MonadTime m) => RunModel WorldState (StateT (Nodes m) m)

sendsInput :: Monad m => Party -> ClientInput Tx -> StateT (Nodes m) m ()

seedWorld :: (MonadDelay m, MonadAsync m, MonadCatch m, MonadTime m) => [(SigningKey HydraKey, b)] -> StateT (Nodes m) m ()

performCommit :: (MonadThrow m, MonadAsync m, MonadTimer m) => [CardanoSigningKey] -> Party -> [(CardanoSigningKey, Value)] -> StateT (Nodes m) m ActualCommitted

performNewTx :: (MonadThrow m, MonadAsync m, MonadTimer m) => Party -> Payment -> StateT (Nodes m) m ()

waitForUTxOToSpend :: (MonadDelay m, MonadThrow m) => UTxO -> CardanoSigningKey -> Value -> Party -> TestHydraNode Tx m -> Int -> StateT (Nodes m) m (Either UTxO (TxIn, TxOut CtxUTxO))

Generator Helpers

genPayment :: WorldState -> Gen (Party, Payment)

genAdaValue :: Gen Value

unsafeConstructorName :: Show a => a -> String

partyKeys :: Gen [(SigningKey HydraKey, CardanoSigningKey)]

Generate a list of pairs of Hydra/Cardano signing keys. All the keys in this list are guaranteed to be unique.

isOwned :: CardanoSigningKey -> (TxIn, TxOut ctx) -> Bool

mkMockTxIn :: VerificationKey PaymentKey -> Word -> TxIn

Orphan instances

Arbitrary Value

Instance details

Methods

arbitrary :: Gen Value

shrink :: Value -> [Value]