Hydra.ModelSpec

Description

Model-Based testing of Hydra Head protocol implementation.

When a property fails it will dump the sequence of actions leading to the failure. This sequence can be copy/pasted and reused directly as a test against either the Model or the implementation as exemplified by the following sample:

@@ it "runs actions against actual nodes" $ do let Actions act = Actions [ Var 1 := Seed { seedKeys = [ (HydraSigningKey (SignKeyEd25519DSIGN "00000000000000000000000000000000000000000000000000000000000000003b6a27bcceb6a42d62a3a8d02a6f0d73653215771de243a63ac048a18b59da29"), "0100000008030606080507030707000607020508050000020207070508040800") , (HydraSigningKey (SignKeyEd25519DSIGN "2e00000000000000000000000000000000000000000000000000000000000000264a0707979e0d6691f74b055429b5f318d39c2883bb509310b67424252e9ef2"), "0106010101070600040403010600080805020003040508030307080706060608") , (HydraSigningKey (SignKeyEd25519DSIGN "ed785af0fb0000000000000000000000000000000000000000000000000000001c02babf6d3d51b725db8b72043823d66634b39db74836b1494bdb647073d566"), "0000070304040705060101030802010105080806050605070104030603010503") ] } , Var 2 := Command{Model.party = Party{vkey = HydraVerificationKey (VerKeyEd25519DSIGN "3b6a27bcceb6a42d62a3a8d02a6f0d73653215771de243a63ac048a18b59da29")}, command = Init{contestationPeriod = -6.413670805613}} , Var 3 := Command{Model.party = Party{vkey = HydraVerificationKey (VerKeyEd25519DSIGN "264a0707979e0d6691f74b055429b5f318d39c2883bb509310b67424252e9ef2")}, command = Commit{Input.utxo = [("0106010101070600040403010600080805020003040508030307080706060608", valueFromList [(AdaAssetId, 18470954)])]}} , Var 4 := Command{Model.party = Party{vkey = HydraVerificationKey (VerKeyEd25519DSIGN "1c02babf6d3d51b725db8b72043823d66634b39db74836b1494bdb647073d566")}, command = Commit{Input.utxo = [("0000070304040705060101030802010105080806050605070104030603010503", valueFromList [(AdaAssetId, 19691416)])]}} , Var 5 := Command{Model.party = Party{vkey = HydraVerificationKey (VerKeyEd25519DSIGN "3b6a27bcceb6a42d62a3a8d02a6f0d73653215771de243a63ac048a18b59da29")}, command = Commit{Input.utxo = [("0100000008030606080507030707000607020508050000020207070508040800", valueFromList [(AdaAssetId, 7003529)])]}} , Var 6 := Command { Model.party = Party{vkey = HydraVerificationKey (VerKeyEd25519DSIGN "3b6a27bcceb6a42d62a3a8d02a6f0d73653215771de243a63ac048a18b59da29")} , command = NewTx { Input.transaction = Payment { from = "0100000008030606080507030707000607020508050000020207070508040800" , to = "0106010101070600040403010600080805020003040508030307080706060608" , value = valueFromList [(AdaAssetId, 7003529)] } } } ] -- env and model state are unused in perform env = []

dummyState :: WorldState (IOSim s) dummyState = WorldState{hydraParties = mempty, hydraState = Start}

loop [] = pure () loop ((Var{} := a) : as) = do void $ perform dummyState a (lookUpVar env) loop as tr = runSimTrace $ evalStateT (loop act) (Nodes mempty traceInIOSim) traceDump = printTrace (Proxy :: Proxy Tx) tr print traceDump True shouldBe True

Documentation

spec :: Spec

prop_generateTraces :: AnyActions -> Property

prop_checkModel :: AnyActions -> Property

assertNodeSeesAndReportsAllExpectedCommits :: GlobalState -> Map Party (TestHydraNode Tx (IOSim s)) -> Party -> PropertyM (StateT (Nodes (IOSim s)) (IOSim s)) ()

assertBalancesInOpenHeadAreConsistent :: GlobalState -> Map Party (TestHydraNode Tx (IOSim s)) -> Party -> PropertyM (StateT (Nodes (IOSim s)) (IOSim s)) ()

waitUntilTheEndOfTime :: MonadDelay m => m ()

Utilities for IOSim

runIOSimProp :: (forall s. Gen (StateT (Nodes (IOSim s)) (IOSim s) Property)) -> Gen Property

Specialised runner similar to https://hackage.haskell.org/package/QuickCheck-2.14.2/docs/src/Test.QuickCheck.Monadic.html#runSTGen.

newtype AnyActions

Constructors

AnyActions

Fields

unAnyActions :: forall s. Actions (WorldState (IOSim s))

unwrapAddress :: AddressInEra -> Text